+++ lib/src/delegate.dart
@@ -5,13 +5,12 @@ import 'controller.dart';
 
 /// A [CustomMultiChildLayout] delegate that renders the nodes in the [InfiniteCanvas].
 class InfiniteCanvasNodesDelegate extends MultiChildLayoutDelegate {
-  InfiniteCanvasNodesDelegate(this.controller);
-  final InfiniteCanvasController controller;
-  List<InfiniteCanvasNode> get children => controller.nodes;
+  InfiniteCanvasNodesDelegate(this.nodes);
+  final List<InfiniteCanvasNode> nodes;
 
   @override
   void performLayout(Size size) {
-    for (final widget in children) {
+    for (final widget in nodes) {
       layoutChild(widget, BoxConstraints.tight(widget.size));
       positionChild(widget, widget.offset);
     }