+++ lib/src/edge_renderer.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 
 import 'controller.dart';
+import 'edge.dart';
 import 'inline_painter.dart';
 
 /// A widget that renders all the edges in the [InfiniteCanvas].
@@ -8,9 +9,14 @@ class InfiniteCanvasEdgeRenderer extends StatelessWidget {
   const InfiniteCanvasEdgeRenderer({
     super.key,
     required this.controller,
+    required this.edges,
+    this.linkStart,
+    this.linkEnd,
   });
 
   final InfiniteCanvasController controller;
+  final List<InfiniteCanvasEdge> edges;
+  final Offset? linkStart, linkEnd;
 
   @override
   Widget build(BuildContext context) {
@@ -22,63 +28,97 @@ class InfiniteCanvasEdgeRenderer extends StatelessWidget {
           ..style = PaintingStyle.stroke
           ..strokeWidth = 2,
         builder: (brush, canvas, rect) {
-          for (final edge in controller.edges) {
+          for (final edge in edges) {
             final from =
                 controller.nodes.firstWhere((node) => node.key == edge.from);
             final to =
                 controller.nodes.firstWhere((node) => node.key == edge.to);
-            final fromRect = from.rect;
-            final toRect = to.rect;
-            // Draw bezier curve from fromRect.center to toRect.center
-            final path = Path();
-            path.moveTo(fromRect.center.dx, fromRect.center.dy);
-            path.cubicTo(
-              fromRect.center.dx,
-              fromRect.center.dy,
-              fromRect.center.dx,
-              toRect.center.dy,
-              toRect.center.dx,
-              toRect.center.dy,
+            drawEdge(
+              context,
+              canvas,
+              from.rect.center,
+              to.rect.center,
+              brush,
+              label: edge.label,
             );
-            canvas.drawPath(path, brush);
-            if (edge.label != null) {
-              final textPainter = TextPainter(
-                text: TextSpan(
-                  text: edge.label,
-                  style: Theme.of(context).textTheme.bodySmall?.copyWith(
-                    color: colors.onSurface,
-                    shadows: [
-                      Shadow(
-                        offset: const Offset(0.8, 0.8),
-                        blurRadius: 3,
-                        color: colors.surface,
-                      ),
-                    ],
-                  ),
-                ),
-                textDirection: TextDirection.ltr,
-              )..layout();
-              // Render label on bezier curve
-              Offset textOffset = Offset(
-                (fromRect.center.dx + toRect.center.dx) / 2,
-                (fromRect.center.dy + toRect.center.dy) / 2,
-              );
-              // Center on curve
-              final pathMetrics = path.computeMetrics();
-              final pathMetric = pathMetrics.first;
-              final pathLength = pathMetric.length;
-              final middle = pathMetric.getTangentForOffset(pathLength / 2);
-              textOffset = middle?.position ?? textOffset;
-              // Offset to top left
-              textOffset = textOffset.translate(
-                -textPainter.width / 2,
-                -textPainter.height / 2,
-              );
-              textPainter.paint(canvas, textOffset);
-            }
+          }
+        },
+      ),
+      foregroundPainter: InlinePainter(
+        brush: Paint()
+          ..color = colors.primary
+          ..style = PaintingStyle.stroke
+          ..strokeWidth = 2,
+        builder: (brush, canvas, child) {
+          if (linkStart != null && linkEnd != null) {
+            drawEdge(
+              context,
+              canvas,
+              linkStart!,
+              controller.toLocal(linkEnd!),
+              brush,
+            );
           }
         },
       ),
     );
+  }
+
+  void drawEdge(
+    BuildContext context,
+    Canvas canvas,
+    Offset fromOffset,
+    Offset toOffset,
+    Paint brush, {
+    String? label,
+  }) {
+    final colors = Theme.of(context).colorScheme;
+    // Draw bezier curve from fromRect.center to toRect.center
+    final path = Path();
+    path.moveTo(fromOffset.dx, fromOffset.dy);
+    path.cubicTo(
+      fromOffset.dx,
+      fromOffset.dy,
+      fromOffset.dx,
+      toOffset.dy,
+      toOffset.dx,
+      toOffset.dy,
+    );
+    canvas.drawPath(path, brush);
+    if (label != null) {
+      final textPainter = TextPainter(
+        text: TextSpan(
+          text: label,
+          style: Theme.of(context).textTheme.bodySmall?.copyWith(
+            color: colors.onSurface,
+            shadows: [
+              Shadow(
+                offset: const Offset(0.8, 0.8),
+                blurRadius: 3,
+                color: colors.surface,
+              ),
+            ],
+          ),
+        ),
+        textDirection: TextDirection.ltr,
+      )..layout();
+      // Render label on bezier curve
+      Offset textOffset = Offset(
+        (fromOffset.dx + toOffset.dx) / 2,
+        (fromOffset.dy + toOffset.dy) / 2,
+      );
+      // Center on curve
+      final pathMetrics = path.computeMetrics();
+      final pathMetric = pathMetrics.first;
+      final pathLength = pathMetric.length;
+      final middle = pathMetric.getTangentForOffset(pathLength / 2);
+      textOffset = middle?.position ?? textOffset;
+      // Offset to top left
+      textOffset = textOffset.translate(
+        -textPainter.width / 2,
+        -textPainter.height / 2,
+      );
+      textPainter.paint(canvas, textOffset);
+    }
   }
 }